+++ packages/in_app_purchase/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.1+1
+
+* Remove skipped driver test.
+
 ## 0.2.1
 
 * iOS: Add currencyCode to priceLocale on productDetails.